% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rows.R
\name{add_formatted_rows}
\alias{add_formatted_rows}
\alias{add_complete_row}
\alias{add_total_row}
\alias{add_complete_total_row}
\alias{add_mean_row}
\alias{add_sd_row}
\alias{add_mean_sd_row}
\alias{add_median_row}
\alias{add_med_row}
\alias{add_q1_row}
\alias{add_q3_row}
\alias{add_q1_q3_row}
\alias{add_iqr_row}
\alias{add_iqr_q3_q1_row}
\alias{add_med_q1_q3_row}
\alias{add_count_rows}
\alias{add_proportion_rows}
\alias{add_prop_rows}
\alias{add_count_prop_rows}
\alias{add_percent_rows}
\alias{add_count_percent_rows}
\alias{add_anova_rows}
\alias{add_chisq_rows}
\alias{add_rao_scott_rows}
\title{Add Formatted Output Rows to Crosstab}
\usage{
add_complete_row(ct)

add_total_row(ct)

add_complete_total_row(ct)

add_mean_row(
  ct,
  round_to = MEAN_ROUND_TO,
  anova_markers = FALSE,
  marker_type = NULL,
  superscript = FALSE,
  cutoff = 0.05
)

add_sd_row(ct, round_to = SD_ROUND_TO)

add_mean_sd_row(
  ct,
  round_to = MEAN_SD_ROUND_TO,
  anova_markers = FALSE,
  marker_type = NULL,
  superscript = FALSE,
  cutoff = 0.05
)

add_median_row(ct, round_to = MEDIAN_ROUND_TO)

add_med_row(ct, round_to = MEDIAN_ROUND_TO)

add_q1_row(ct, round_to = Q1_ROUND_TO)

add_q3_row(ct, round_to = Q3_ROUND_TO)

add_q1_q3_row(ct, round_to = Q1_Q3_ROUND_TO)

add_iqr_row(ct, round_to = IQR_ROUND_TO)

add_iqr_q3_q1_row(ct, round_to = IQR_Q3_Q1_ROUND_TO)

add_med_q1_q3_row(ct, round_to = MED_Q1_Q3_ROUND_TO)

add_count_rows(ct, keep_na_vars = FALSE)

add_proportion_rows(ct, round_to = PROP_ROUND_TO, keep_na_vars = FALSE)

add_prop_rows(ct, round_to = PROP_ROUND_TO, keep_na_vars = FALSE)

add_count_prop_rows(ct, round_to = COUNT_PROP_ROUND_TO, keep_na_vars = FALSE)

add_percent_rows(
  ct,
  round_to = PERCENT_ROUND_TO,
  keep_na_vars = FALSE,
  raw = FALSE
)

add_count_percent_rows(
  ct,
  round_to = COUNT_PERCENT_ROUND_TO,
  keep_na_vars = FALSE
)

add_anova_rows(ct, cutoff = 0.05, round_to = 3)

add_chisq_rows(ct, p.adj = TRUE, method = "BH", cutoff = 0.05, round_to = 3)

add_rao_scott_rows(
  ct,
  p.adj = TRUE,
  method = "BH",
  cutoff = 0.05,
  round_to = 3
)
}
\arguments{
\item{ct}{The crosstab object used to calculate the values}

\item{round_to}{Numeric - How many decimal places to round the values to}

\item{anova_markers}{Logical - Append marker characters to indicate significant differences between cohorts}

\item{marker_type}{Character - What the markers should be; "symbol", "number", or "alphabet"}

\item{superscript}{Logical - Should the markers be superscripted in the kable output table? (This will work in final Latex output, but will fail in the Viewer)}

\item{cutoff}{Numeric - The p-value cutoff for significance (usually 0.05)}

\item{keep_na_vars}{Logical - If `TRUE`, NA variables are treated as values and shown in rows like "count"}

\item{raw}{Logical - If `TRUE` the function returns the raw numeric values rather than the formatted strings (e.g. 25 instead of "25\%")}

\item{p.adj}{Logical - Should the p-values be adjusted to account for multiple pairwise tests?}

\item{method}{Character - The p-value adjustment method (see the man page for [stats::p.adjust()] for possible values)}
}
\value{
A crosstab object with the specified row(s) added to it
}
\description{
The `add_*_row()` functions all create a new, formatted row with the
specified data and add it to the crosstab object.
}
\section{Functions}{
\itemize{
\item \code{add_complete_row()}: Add the number complete (not NA) for each cohort

\item \code{add_total_row()}: Add the number of rows (including NA) for each cohort

\item \code{add_complete_total_row()}: Add a string formatted as "[complete] / [total]" for each cohort

\item \code{add_mean_row()}: Add the mean for each cohort

\item \code{add_sd_row()}: Add the standard deviation for each cohort

\item \code{add_mean_sd_row()}: Add a string formatted as "[mean] +/- [sd]" for each cohort

\item \code{add_median_row()}: Add the median for each cohort (same as `get_med_row()`)

\item \code{add_med_row()}: Add the median for each cohort (same as `get_median_row()`)

\item \code{add_q1_row()}: Add the first quartile for each cohort

\item \code{add_q3_row()}: Add the third quartile for each cohort

\item \code{add_q1_q3_row()}: Add a string formatted as "[Q1]--[Q3]" for each cohort

\item \code{add_iqr_row()}: Add the IQR (Q3-Q1) for each cohort

\item \code{add_iqr_q3_q1_row()}: Add a string formatted as "[IQR] ([Q3]-[Q1])" for each cohort

\item \code{add_med_q1_q3_row()}: Add a string formatted as "[median] ([Q1], [Q3])" for each cohort

\item \code{add_count_rows()}: Add the counts for each response and cohort

\item \code{add_proportion_rows()}: Add the proportion of responses (out of complete) for each response and cohort (same as `get_prop_rows()`)

\item \code{add_prop_rows()}: Add the proportion of responses (out of complete) for each response and cohort (same as `get_proportion_rows()`)

\item \code{add_count_prop_rows()}: Add a string formatted as "[count] ([prop])" for each response and cohort

\item \code{add_percent_rows()}: Add the percent (out of complete) for each response and cohort

\item \code{add_count_percent_rows()}: Add a string formatted as "[count] ([percent]\%)" for each response and cohort

\item \code{add_anova_rows()}: Add the rows formatted with ANOVA p-value and Tukey post-hoc (if applicable)

\item \code{add_chisq_rows()}: Add the rows formatted with chi-square p-value and pairwise post-hoc (if applicable)

\item \code{add_rao_scott_rows()}: Add the rows formatted with Rao-Scott chi-square p-value and pairwise posthoc (if applicable)

}}
\section{Row Format Components}{

The rows contain some formatted combination of these variables:
* Total: The number of rows in the original data frame
* Complete: The number of non-NA rows in the original data frame
* Mean: The mean of the numeric data or Likert-like data
* SD: The standard deviation of the numeric data or Likert-like data
* Med: The median of the numeric data
* Q1: The 1st quartile of the numeric data
* Q3: The 3rd quartile of the numeric data
* Count: The total number of each response in the categorical data
* Proportion: The proportion of each response in the categorical data (out of "complete")
* Percent: The percent of each response in the categorical data (out of "complete")
* ANOVA: The ANOVA p-values and Tukey post-hoc p-values of the numeric data
* Chi-Square: The chi-square p-values and adjusted pairwise post-hoc p-values of the numeric data
* Rao-Scott: The Rao-Scott adjusted chi-square p-values and adjusted pairwise post-hoc p-values of the numeric data
}

\examples{
num_df <- iris[, c("Sepal.Length", "Species"), drop = FALSE]
num_ct <- crosstab(num_df, cohort_col_name = "Species")

num_ct # Empty

num_ct <- num_ct |>
    add_complete_total_row() |>
    add_mean_sd_row(anova_markers = TRUE) |>
    add_med_q1_q3_row() |>
    add_anova_rows()

num_ct # Filled with several rows

}
