% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{add_formatted_table}
\alias{add_formatted_table}
\alias{add_categorical_table}
\alias{add_numeric_table}
\alias{add_likert_table}
\alias{add_default_table}
\title{Add Pre-Built Tables}
\usage{
add_categorical_table(
  ct,
  chisq = TRUE,
  keep_na_vars = FALSE,
  round_percent_to = PERCENT_ROUND_TO,
  p.adj = TRUE,
  method = "BH",
  cutoff = 0.05,
  round_p_val_to = 3
)

add_numeric_table(
  ct,
  anova = TRUE,
  round_mean_sd_to = MEAN_SD_ROUND_TO,
  round_med_q1_q3_to = MED_Q1_Q3_ROUND_TO,
  round_p_val_to = 3,
  cutoff = 0.05,
  anova_format = "row",
  marker_type = NULL,
  superscript = FALSE
)

add_likert_table(
  ct,
  anova = TRUE,
  chisq = TRUE,
  keep_na_vars = FALSE,
  round_mean_sd_to = MEAN_SD_ROUND_TO,
  round_percent_to = PERCENT_ROUND_TO,
  round_p_val_to = 3,
  p.adj = TRUE,
  method = "BH",
  cutoff = 0.05,
  anova_format = "row",
  marker_type = NULL,
  superscript = FALSE
)

add_default_table(
  ct,
  anova = TRUE,
  chisq = TRUE,
  keep_na_vars = FALSE,
  round_mean_sd_to = MEAN_SD_ROUND_TO,
  round_med_q1_q3_to = MED_Q1_Q3_ROUND_TO,
  round_percent_to = PERCENT_ROUND_TO,
  round_p_val_to = 3,
  p.adj = TRUE,
  method = "BH",
  cutoff = 0.05,
  anova_format = "row",
  marker_type = NULL,
  superscript = FALSE
)
}
\arguments{
\item{ct}{The crosstab object to add to}

\item{chisq}{Logical - Should chi-square tests be performed and info added?}

\item{keep_na_vars}{Logical - Should NA variables be kept in the output of "count" rows?}

\item{round_percent_to}{Numeric - How many digits to round the percent to}

\item{p.adj}{Logical - Should the chi-square post-hoc p-values be adjusted to account for multiple tests?}

\item{method}{Character - The p-value adjustment method (see the man page for [stats::p.adjust()] for possible values)}

\item{cutoff}{Numeric - The p-value cutoff for significance (usually 0.05)}

\item{round_p_val_to}{Numeric - How many digits to round the p-value to}

\item{anova}{Logical - Should ANOVA tests be performed and info added?}

\item{round_mean_sd_to}{Numeric - How many digits to round the mean and sd to}

\item{round_med_q1_q3_to}{Numeric - How many digits to round the median, q1, and q3 to}

\item{anova_format}{Character - "row" for ANOVA data as rows, "marker" for ANOVA data as markers appended to mean and sd data, and c("row", "marker") for both}

\item{marker_type}{Character - What the markers should be; "symbol", "number", or "alphabet"}

\item{superscript}{Logical - Should the markers be superscripted in the kable output table? (This will work in final Latex output, but will fail in the Viewer)}
}
\value{
The crosstab object with the specified tables added to it
}
\description{
The `add_*_table()` functions add pre-built sets of rows for different types
of data (e.g. numeric, Likert-like, and categorical - including multi-response).
`add_default_table()` automatically detects what kind of table to add based
on the data type.
}
\section{Functions}{
\itemize{
\item \code{add_categorical_table()}: Add the default table for categorical data (including multi-response)

\item \code{add_numeric_table()}: Add the default table for numeric data

\item \code{add_likert_table()}: Add the default table for Likert-like data

\item \code{add_default_table()}: Add the proper table based on data type

}}
\section{Categorical Table}{

* "Complete / Total" row
* Count rows
* Chi-square rows (unless chisq = FALSE)
}

\section{Numeric Table}{

* "Complete / Total" row
* "Mean +/- SD" row
* "Median (Q1, Q3)" row
* ANOVA rows (unless anova = FALSE)
}

\section{Likert Table}{

* "Complete / Total" row
* "Mean +/- SD" row
* Count rows
* ANOVA rows (unless anova = FALSE)
* Chi-square rows (unless chisq = FALSE)
}

\examples{
# Numeric Table
num_ct <- crosstab(
    length_by_species,
    cohort_col_name = "species"
)

num_ct <- num_ct |>
    add_numeric_table()

num_ct

# Likert table
licorice_map <- c("likes" = 1, "neither" = 0, "dislikes" = -1)
likert_ct <- crosstab(
    licorice_by_region,
    cohort_col_name = "region",
    var_map = licorice_map
)

likert_ct <- likert_ct |>
    add_likert_table()

likert_ct
}
