% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rows.R
\name{add_rows}
\alias{add_rows}
\title{Add Rows to A Crosstab Object}
\usage{
add_rows(ct, rows, ind = NULL, index_from = "top", table_name = NULL)
}
\arguments{
\item{ct}{The crosstab to add to}

\item{rows}{The rows to add to the crosstab}

\item{ind}{(Optional) The specific row index at which to add the row}

\item{index_from}{(Optional) Either "top" or "bottom", whether to count the previously specified index from the top or bottom of the table}

\item{table_name}{Pass a specific table name to tell `index()` that this is a new group of rows}
}
\value{
The crosstab passed in, but with the extra rows added and attributes updated.
}
\description{
Add a custom or pre-built row (like the `get_*_row()` functions) to the
crosstab object and update all relevant attributes.
}
\details{
You can insert the rows at a specific place on the table, counting from the
top or bottom. The default is inserting at the end of the table, but setting
`index_from = "bottom"` will insert it at the beginning (last from the
bottom). `index = 3` will insert it before row 3 so it becomes row 3, and
`index = 3, index_from = "bottom"` will insert it 3rd from the bottom.
}
\examples{
num_ct <- crosstab(length_by_species, cohort_col_name = "species")

complete_total_row <- get_complete_total_row(num_ct)
complete_total_row

mean_sd_row <- get_mean_sd_row(num_ct)
mean_sd_row

num_ct <- num_ct |>
    add_rows(complete_total_row) |>
    add_rows(mean_sd_row)
num_ct

}
