% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{allergies_by_school}
\alias{allergies_by_school}
\title{Allergies by High School}
\format{
A data frame with 400 rows of multi-response allergy information. If no allergies, the item is a zero-length character vector.
\describe{
  \item{allergies}{Allergy lists (list of factors)}
  \item{school}{High School: Brighton High, Cottonwood High, Olympus High, and Alta High (factor)}
}
}
\source{
Dummy data, invented by the developer
}
\usage{
allergies_by_school
}
\description{
A dummy dataset of multi-response (select all that apply) data concerning
allergies of students at local high schools.
}
\keyword{datasets}
