% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_extensions.R
\name{apply_footnotes}
\alias{apply_footnotes}
\title{Apply Auto-Generated Footnotes to Kable Objects from Crosstabs}
\usage{
apply_footnotes(kable_input, ct)
}
\arguments{
\item{kable_input}{A kable object created by `kbl()`}

\item{ct}{The associated crosstab used to create the kable object}
}
\value{
The kable object with the footnotes added
}
\description{
`apply_footnotes()` takes in kable objects created by [kbl()] and uses
[kableExtra::footnote()] to apply auto-generated footnotes (e.g. Likert maps,
legends for ANOVA markers, or extra info about statistical tests). Unless
specified otherwise, this function is automatically called by
[htcrosstabs::kbl()].
}
\examples{

test_df <- iris[, c("Petal.Length", "Species"), drop = FALSE]
test_ct <- crosstab(test_df, "Species") |>
    add_default_table(anova_format = c("row", "marker"))

printable_output <- kbl(test_ct, crosstab_footnotes = FALSE)
printable_output # Without footnotes

printable_output <- apply_footnotes(printable_output, test_ct)
printable_output # With footnotes

}
