% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_extensions.R
\name{apply_pack_rows}
\alias{apply_pack_rows}
\title{Apply Auto-Generated Row Packing to Kable Objects from Crosstabs}
\usage{
apply_pack_rows(kable_input, ct)
}
\arguments{
\item{kable_input}{A kable object created by `kbl()`}

\item{ct}{The associated crosstab used to create the kable object}
}
\value{
The kable object with the row grouping added
}
\description{
`apply_pack_rows()` takes in kable objects created by [kbl()] and uses
[kableExtra::pack_rows()] to group the rows into sections with an
auto-generated header (pulled from `index()`). Unless specified otherwise,
this function is automatically called by [htcrosstabs::kbl()].
}
\examples{
test_df <- iris[, c("Petal.Length", "Species"), drop = FALSE]
test_ct <- crosstab(test_df, "Species") |>
    add_default_table(anova_format = c("row", "marker"))

printable_output <- kbl(test_ct, pack_rows = FALSE)
printable_output # Without row grouping

index(test_ct) # The index used for grouping

printable_output <- apply_pack_rows(printable_output, test_ct)
printable_output # With row grouping

}
