% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_extensions.R
\name{apply_table_sep}
\alias{apply_table_sep}
\title{Apply Lines Beetween Stacked Tables in Kable Objects from Crosstabs}
\usage{
apply_table_sep(kable_input, ct)
}
\arguments{
\item{kable_input}{A kable object created by `kbl()`}

\item{ct}{The associated crosstab used to create the kable object}
}
\value{
The kable object with the table-separating lines added
}
\description{
`apply_table_sep()` takes in kable objects created by [kbl()] and uses
[kableExtra::row_spec()] to apply horizontal lines between stacked tables.
Unless specified otherwise, this function is automatically called by
[htcrosstabs::kbl()].
}
\examples{
test_ct <- crosstab_stacked(iris, "Species")

printable_output <- kbl(test_ct, table_sep = FALSE)
printable_output # Without lines

printable_output <- apply_table_sep(printable_output, test_ct)
printable_output # With lines (likely won't be visible in Viewer window)

}
