% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_casting.R
\name{cast_crosstab}
\alias{cast_crosstab}
\alias{as.crosstab.cat}
\alias{as.crosstab.num}
\alias{as.crosstab.likert}
\alias{as.crosstab.multi}
\title{Cast Crosstab Object to New Data Type}
\usage{
as.crosstab.cat(ct_data)

as.crosstab.num(ct_data)

as.crosstab.likert(ct_data, var_map = NULL)

as.crosstab.multi(ct_data)
}
\arguments{
\item{ct_data}{The data or crosstab object to be cast}

\item{var_map}{A named numeric vector mapping character values to numeric values}
}
\value{
The converted crosstab object
}
\description{
`as.crosstab.*()` functions take preexisting data and convert it to
a new type of crosstab object (e.g. numeric -> categorical, likert ->
numeric, etc.)
}
\section{Functions}{
\itemize{
\item \code{as.crosstab.cat()}: Casts the crosstab object to categorical

\item \code{as.crosstab.num()}: Casts the crosstab object to numeric

\item \code{as.crosstab.likert()}: Casts the crosstab object to Likert-like

\item \code{as.crosstab.multi()}: Casts the crosstab object to multi-answer

}}
\section{Converting to Categorical}{

Numeric values are simply cast using `as.factor()`, internally storing them
as their character representations.

Likert-like data is already categorical, but with extra mapping
functionality, bestowed by `var_map`. Mapping functionality is dropped when
casting.

Multi-response data simply has its list-column unnested into one answer per
row.
}

\section{Converting to Numeric}{

Categorical values are cast using `as.numeric()`, which replaces the data
with its numeric factor level.

Likert-like data has `var_map` for this exact purpose. The character values
are found in the names of `var_map` and then mapped to their corresponding
numeric values.

Multi-response data is unnested to one answer per row, and then undergoes
the same process as categorical data.
}

\section{Converting to Likert-Like}{

Categorical values are kept, and the provided `var_map` is attached to the
crosstab object.

Numeric data is reverse-mapped to categorical values by their values in
`var_map`.

Multi-response data is unnested to one response per row and then the same
as categorical values.
}

\section{Converting to Multi-Response}{

Categorical values are turned into a list-column where each row still has
one answer per row.

Numeric values are cast using `as.factor()` and then the same as categorical
values.

Likert-like data has "mapping" functionality dropped and then the same as
categorical values.
}

\examples{
num_ct <- crosstab(length_by_species, "species")
converted_ct <- as.crosstab.cat(num_ct)
is.crosstab.categorical(converted_ct) # TRUE

# Likert data requires a map
cat_ct <- crosstab(licorice_by_region, "region")
licorice_map <- c("likes" = 1, "neither" = 0, "dislikes" = -1)
converted_ct <- as.crosstab.likert(cat_ct, var_map = licorice_map)
is.crosstab.likert(converted_ct) # TRUE
}
