% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_attributes.R
\name{class_attributes}
\alias{class_attributes}
\alias{var_name}
\alias{var}
\alias{var_levels}
\alias{cohort_name}
\alias{cohort}
\alias{cohort_levels}
\alias{var_map}
\alias{var_mapped}
\alias{combined_cohort_name}
\alias{get_raw_data}
\alias{desc_name}
\alias{var_name<-}
\alias{var<-}
\alias{var_levels<-}
\alias{cohort_name<-}
\alias{cohort<-}
\alias{cohort_levels<-}
\alias{var_map<-}
\title{Get Values from Stored Data}
\usage{
var_name(ct_data)

var(ct_data, raw = FALSE)

var_levels(ct_data)

cohort_name(ct_data)

cohort(ct_data, raw = FALSE)

cohort_levels(ct_data, raw = FALSE)

var_map(ct_data)

var_mapped(ct_data, raw = FALSE)

combined_cohort_name(ct_data)

get_raw_data(ct_data)

desc_name(ct_data)

var_name(ct_data) <- value

var(ct_data) <- value

var_levels(ct_data) <- value

cohort_name(ct_data) <- value

cohort(ct_data) <- value

cohort_levels(ct_data) <- value

var_map(ct_data) <- value
}
\arguments{
\item{ct_data}{The crosstab or crosstab_data object you want to extract the data from}

\item{raw}{Logical, should the combined or "All" cohort data be removed?}

\item{value}{The replacement value}
}
\value{
The specified data
}
\description{
The getter functions will return different bits of data or return the data
in different formats as needed. They can be called on both crosstab and
crosstab_data objects.

The setter functions will set those same values if provided with valid data.
}
\section{Functions}{
\itemize{
\item \code{var_name()}: Returns the name of the variable column

\item \code{var()}: Returns the variable column as a vector

\item \code{var_levels()}: Returns the factor levels in the variable column

\item \code{cohort_name()}: Returns the name of the cohort column

\item \code{cohort()}: Returns the cohort column as a vector

\item \code{cohort_levels()}: Returns the factor levels in the cohort column

\item \code{var_map()}: Returns the named numeric vector used to map likert values to their numeric value

\item \code{var_mapped()}: Returns the variable column, except all values are mapped using var_map

\item \code{combined_cohort_name()}: Returns the name of the combined cohort added to grouped data

\item \code{get_raw_data()}: Returns the crosstab_data object without the extra combined cohort data

\item \code{desc_name()}: Returns the name of the leftmost description column in the output

\item \code{var_name(ct_data) <- value}: Sets the name of the variable column

\item \code{var(ct_data) <- value}: Sets the values in the variable column

\item \code{var_levels(ct_data) <- value}: Sets the factor levels in the variable column

\item \code{cohort_name(ct_data) <- value}: Sets the name of the cohort column

\item \code{cohort(ct_data) <- value}: Sets the values in the cohort column

\item \code{cohort_levels(ct_data) <- value}: Sets the factor levels in the cohort column

\item \code{var_map(ct_data) <- value}: Sets the numeric values to map the character variables to

}}
\examples{
test_ct <- crosstab(
    length_by_species,
    "species",
    new_var_col_name = "orig_var_name"
)

var_name(test_ct) # orig_var_name
var_name(test_ct) <- "new_var_name"
var_name(test_ct) # new_var_name

}
