% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables.R
\name{crosstab_stacked}
\alias{crosstab_stacked}
\title{Create Stacked Tables From Multiple Data Columns}
\usage{
crosstab_stacked(
  df,
  cohort_col_name = NULL,
  var_map = NULL,
  combined_cohort_name = "All",
  desc_col_name = "Description",
  anova = TRUE,
  chisq = TRUE,
  keep_na_vars = FALSE,
  round_mean_sd_to = MEAN_SD_ROUND_TO,
  round_med_q1_q3_to = MED_Q1_Q3_ROUND_TO,
  round_percent_to = PERCENT_ROUND_TO,
  round_p_val_to = 3,
  p.adj = TRUE,
  method = "BH",
  cutoff = 0.05,
  anova_format = "row",
  marker_type = NULL,
  superscript = FALSE
)
}
\arguments{
\item{df}{The original data frame with the data to be analyzed}

\item{cohort_col_name}{The name of the cohort column, so it can be paired with each sub-crosstab}

\item{var_map}{A named numeric vector (or list of vectors) for Likert-like data mapping categorical values to numeric values}

\item{combined_cohort_name}{The name for the combined or "All" column for grouped data}

\item{desc_col_name}{The name for the leftmost "description" column in the output table}

\item{anova}{Logical - Should ANOVA tests be performed and info added?}

\item{chisq}{Logical - Should chi-square tests be performed and info added?}

\item{keep_na_vars}{Logical - Should NA variables be kept in the output of "count" rows?}

\item{round_mean_sd_to}{Numeric - How many digits to round the mean and sd to}

\item{round_med_q1_q3_to}{Numeric - How many digits to round the median, q1, and q3 to}

\item{round_percent_to}{Numeric - How many digits to round the percent to}

\item{round_p_val_to}{Numeric - How many digits to round the p-value to}

\item{p.adj}{Logical - Should the chi-square post-hoc p-values be adjusted to account for multiple tests?}

\item{method}{Character - The p-value adjustment method (see the man page for [stats::p.adjust()] for possible values)}

\item{cutoff}{Numeric - The p-value cutoff for significance (usually 0.05)}

\item{anova_format}{Character - "row" for ANOVA data as rows, "marker" for ANOVA data as markers appended to mean and sd data, and c("row", "marker") for both}

\item{marker_type}{Character - What the markers should be; "symbol", "number", or "alphabet"}

\item{superscript}{Logical - Should the markers be superscripted in the kable output table? (This will work in final Latex output, but will fail in the Viewer)}
}
\value{
The stacked crosstab object
}
\description{
While [crosstab()] handles the base case of one variable column (and one
cohort column if applicable), `crosstab_stacked()` iterates over each
variable column to create a stacked crosstab object with multiple combined
tables. It uses the same arguments that [crosstab()] does to determine type.
If you have 2 likert columns with different var_maps, you can make a list
with both and pass that in (the names of the list should match the columns
they correspond to).
}
\examples{
# All numeric data
crosstab_stacked(
    iris,
    cohort_col_name = "Species"
)

# Combining different types of data
uni_perception_map <- c(
    "overwhelmingly positive" = 2,
    "somewhat positive" = 1,
    "neutral" = 0,
    "somewhat negative" = -1,
    "overwhelmingly negative" = -2
)

prof_support_map <- c(
    "very supportive" = 4,
    "somewhat supportive" = 3,
    "apathetic" = 2,
    "somewhat demeaning" = 1,
    "very demeaning" = 0
)

likert_maps <- list(
    uni_perception = uni_perception_map,
    prof_support = prof_support_map
)

crosstab_stacked(
    students,
    cohort_col_name = "university",
    var_map = likert_maps,
    anova = FALSE,
    chisq = FALSE
)

}
