% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{default_var_map}
\alias{default_var_map}
\title{Create Default Variable Map from Factor Levels}
\usage{
default_var_map(fct)
}
\arguments{
\item{fct}{A factor or factorlist}
}
\value{
A named numeric vector with the factors mapped as names and their scores as values
}
\description{
Takes a factor or factorlist and assigns them Likert scores from
`length(fct)` to 1 in descending order, so a vector with the levels
`c("A", "B", "C")` would return `c(A = 3, B = 2, C = 1)`.
}
\examples{
levels(students$prof_support)
default_var_map(students$prof_support)

}
