% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_p_values}
\alias{get_p_values}
\alias{get_anova_p_value}
\alias{get_chisq_p_value}
\alias{get_rao_scott_p_value}
\title{Perform Statistical Tests on Crosstab Objects}
\usage{
get_anova_p_value(data)

get_chisq_p_value(data)

get_rao_scott_p_value(data)
}
\arguments{
\item{data}{The crosstab or crosstab_data object}
}
\value{
The specified p-value
}
\description{
`get_*_p_value()` functions take crosstab objects and return a single p-value
based on the given test: ANOVA for numeric and Likert-like data, chi-squared
categorical and Likert-like, and Rao-Scott adjusted chi-square for
multiresponse data. Post-hoc tests are performed by the [get_*_posthoc()]
functions.
}
\section{Functions}{
\itemize{
\item \code{get_anova_p_value()}: Perform ANOVA test

\item \code{get_chisq_p_value()}: Perform chi-square test

\item \code{get_rao_scott_p_value()}: Perform Rao-Scott corrected chi-square test

}}
\section{ANOVA Tests}{

ANOVA tests are designed for numeric data and Likert-like data that has been
mapped to numeric. It can be performed on the other data types, but they are
cast first (also there would be very little info gained). The ANOVA test is
performed using [stats::aov()].
}

\section{Chi-Square Tests}{

Chi-square tests are designed for categorical data and Likert-like data
(unmapped). Multi-response data is passed to the Rao-Scott adjusted
chi-square test with `get_rao_scott_p_value()`. It can be performed on
numeric data, but it is cast first (and there would be little information to
be gained). The chi-square test is performed using [stats::chisq.test()].
}

\section{Rao-Scott Adjusted Chi-Square Tests}{

Rao-Scott adjusted chi-square tests are a corrected version of chi-square
tests used for multi-response data. It can be used on other types of
responses, but they are cast first, and the response will be essentially the
same as the normal chi-square test. The Rao-Scott adjusted test is
performed using [survey::svychisq()].
}

\examples{
# ANOVA test for numeric data
num_df <- iris[, c("Sepal.Length", "Species"), drop = FALSE]
num_ct <- crosstab(num_df, "Species")

get_anova_p_value(num_ct)

# Chi-square test for categorical data
cat_df <- data.frame(
    Age = sample(c("Child", "Teen", "Adult"), 200, replace = TRUE),
    Sport = sample(c("Soccer", "Basketball", "Football"), 200, replace = TRUE)
)
cat_ct <- crosstab(cat_df, "Age")

get_chisq_p_value(cat_ct)

}
