% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{get_posthoc}
\alias{get_posthoc}
\alias{get_tukey_posthoc}
\alias{get_anova_posthoc}
\alias{get_chisq_posthoc}
\alias{get_rao_scott_posthoc}
\title{Perform Post-Hoc Tests on Crosstab Objects}
\usage{
get_tukey_posthoc(data)

get_anova_posthoc(data)

get_chisq_posthoc(data, p.adj = TRUE, method = "BH")

get_rao_scott_posthoc(data, p.adj = TRUE, method = "BH")
}
\arguments{
\item{data}{The crosstab or crosstab_data object}

\item{p.adj}{Logical - Should the p-values be adjusted to correct for multiple tests?}

\item{method}{Character - The p-value adjustment method (see the man page for [stats::p.adjust()] for possible values)}
}
\value{
A symmetric data frame/matrix with pairwise post-hoc p-values
}
\description{
`get_*_posthoc()` functions take crosstab objects and return a symmetric
data frame/matrix of pairwise comparison p-values based on the given test:
ANOVA for numeric and Likert-like data, chi-squared categorical and
Likert-like, and Rao-Scott adjusted chi-square for multiresponse data.
Overall statistical tests to get a single p-value for all data collectively
are performed by the [get_*_p_value()] functions.

The p-values from the multiple comparisons can be corrected using methods
such as Bonferroni, Benjamini-Hochberg, etc. using the `method` parameter.
P-values are adjusted using Benjamini Hochberg by default, but if no
adjustment is desired, set `p.adj = FALSE`.
}
\section{Functions}{
\itemize{
\item \code{get_tukey_posthoc()}: Perform Tukey post-hoc

\item \code{get_anova_posthoc()}: Perform Tukey post-hoc (same as `get_tukey_posthoc()`)

\item \code{get_chisq_posthoc()}: Perform chi-square post-hoc

\item \code{get_rao_scott_posthoc()}: Perform Rao-Scott corrected chi-square post-hoc

}}
\section{ANOVA Tests}{

ANOVA tests are designed for numeric data and Likert-like data that has been
mapped to numeric. It can be performed on the other data types, but they are
cast first (also there would be very little info gained). The ANOVA test is
performed using [stats::aov()].
}

\section{Chi-Square Tests}{

Chi-square tests are designed for categorical data and Likert-like data
(unmapped). Multi-response data is passed to the Rao-Scott adjusted
chi-square test with `get_rao_scott_p_value()`. It can be performed on
numeric data, but it is cast first (and there would be little information to
be gained). The chi-square test is performed using [stats::chisq.test()].
}

\section{Rao-Scott Adjusted Chi-Square Tests}{

Rao-Scott adjusted chi-square tests are a corrected version of chi-square
tests used for multi-response data. It can be used on other types of
responses, but they are cast first, and the response will be essentially the
same as the normal chi-square test. The Rao-Scott adjusted test is
performed using [survey::svychisq()].
}

\examples{
# ANOVA test for numeric data
num_df <- iris[, c("Sepal.Length", "Species"), drop = FALSE]
num_ct <- crosstab(num_df, "Species")

get_tukey_posthoc(num_ct)

# Chi-square test for categorical data
cat_df <- data.frame(
    Age = sample(c("Child", "Teen", "Adult"), 200, replace = TRUE),
    Sport = sample(c("Soccer", "Basketball", "Football"), 200, replace = TRUE)
)
cat_ct <- crosstab(cat_df, "Age")

get_chisq_posthoc(cat_ct)

}
