% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vals.R
\name{get_values}
\alias{get_values}
\alias{get_complete}
\alias{get_total}
\alias{get_complete_total}
\alias{get_mean}
\alias{get_sd}
\alias{get_mean_sd}
\alias{get_median}
\alias{get_med}
\alias{get_q1}
\alias{get_q3}
\alias{get_q1_q3}
\alias{get_iqr}
\alias{get_iqr_q3_q1}
\alias{get_med_q1_q3}
\alias{get_count}
\alias{get_proportion}
\alias{get_prop}
\alias{get_count_prop}
\alias{get_percent}
\alias{get_count_percent}
\title{Get Values Calculated From Crosstab}
\usage{
get_complete(ct_data, out_col_name = COMP_COL_NAME)

get_total(ct_data, out_col_name = TOTAL_COL_NAME)

get_complete_total(ct_data, out_col_name = COMP_TOT_COL_NAME, raw = FALSE)

get_mean(ct_data, out_col_name = MEAN_COL_NAME, round_to = MEAN_ROUND_TO)

get_sd(ct_data, out_col_name = SD_COL_NAME, round_to = SD_ROUND_TO)

get_mean_sd(
  ct_data,
  out_col_name = MEAN_SD_COL_NAME,
  round_to = MEAN_SD_ROUND_TO,
  raw = FALSE
)

get_median(ct_data, out_col_name = MED_COL_NAME, round_to = MEDIAN_ROUND_TO)

get_med(ct_data, out_col_name = MED_COL_NAME, round_to = MEDIAN_ROUND_TO)

get_q1(ct_data, out_col_name = Q1_COL_NAME, round_to = Q1_ROUND_TO)

get_q3(ct_data, out_col_name = Q3_COL_NAME, round_to = Q3_ROUND_TO)

get_q1_q3(
  ct_data,
  out_col_name = Q1_Q3_COL_NAME,
  round_to = Q1_Q3_ROUND_TO,
  raw = FALSE
)

get_iqr(ct_data, out_col_name = IQR_COL_NAME, round_to = IQR_ROUND_TO)

get_iqr_q3_q1(
  ct_data,
  out_col_name = IQR_Q3_Q1_COL_NAME,
  round_to = IQR_Q3_Q1_ROUND_TO,
  raw = FALSE
)

get_med_q1_q3(
  ct_data,
  out_col_name = MED_Q1_Q3_COL_NAME,
  round_to = MED_Q1_Q3_ROUND_TO,
  raw = FALSE
)

get_count(ct_data, out_col_name = COUNT_COL_NAME, keep_na_vars = FALSE)

get_proportion(
  ct_data,
  out_col_name = PROP_COL_NAME,
  round_to = PROP_ROUND_TO,
  keep_na_vars = FALSE
)

get_prop(
  ct_data,
  out_col_name = PROP_COL_NAME,
  round_to = PROP_ROUND_TO,
  keep_na_vars = FALSE
)

get_count_prop(
  ct_data,
  out_col_name = COUNT_PROP_COL_NAME,
  round_to = COUNT_PROP_ROUND_TO,
  keep_na_vars = FALSE,
  raw = FALSE
)

get_percent(
  ct_data,
  out_col_name = PERCENT_COL_NAME,
  round_to = PERCENT_ROUND_TO,
  keep_na_vars = FALSE,
  raw = FALSE
)

get_count_percent(
  ct_data,
  out_col_name = COUNT_PERCENT_COL_NAME,
  round_to = COUNT_PERCENT_ROUND_TO,
  keep_na_vars = FALSE,
  raw = FALSE
)
}
\arguments{
\item{ct_data}{The crosstab or crosstab_data object}

\item{out_col_name}{Character - A custom name for the variable column in the output}

\item{raw}{Logical - If `TRUE` the data frame returned contains the raw values rather than one formatted column}

\item{round_to}{Numeric - The number of digits to round the specified value to}

\item{keep_na_vars}{Logical - Whether or not NA values should be included in variables like count}
}
\value{
A data frame with the specified data as well as other columns to differentiate
}
\description{
The `get_*()` functions return a data frame with the specified data
calculated from the provided crosstab, casted if necessary. These are the
base values that the other functions use to produce their formatted output.
}
\section{Functions}{
\itemize{
\item \code{get_complete()}: Get the number of non-NA rows in the data

\item \code{get_total()}: Get the number of rows in the data (including NA)

\item \code{get_complete_total()}: Get a string formatted as "[complete] / [total]" for each cohort

\item \code{get_mean()}: Get the means for each cohort

\item \code{get_sd()}: Get the standard deviations for each cohort

\item \code{get_mean_sd()}: Get a string formatted as "[mean] +/- [sd]" for each cohort

\item \code{get_median()}: Get the median for each cohort (same as `get_med()`)

\item \code{get_med()}: Get the median for each cohort (same as `get_median()`)

\item \code{get_q1()}: Get the first quartile for each cohort

\item \code{get_q3()}: Get the third quartile for each cohort

\item \code{get_q1_q3()}: Get a string formatted as "[Q1]--[Q3]" for each cohort

\item \code{get_iqr()}: Get the IQR for each cohort

\item \code{get_iqr_q3_q1()}: Get a string formatted as "[IQR] ([Q3]-[Q1])" for each cohort

\item \code{get_med_q1_q3()}: Get a string formatted as "[median] ([Q1], [Q3])" for each cohort

\item \code{get_count()}: Get the counts for each variable and cohort

\item \code{get_proportion()}: Get the proportions (out of complete) for each variable and cohort (same as `get_prop()`)

\item \code{get_prop()}: Get the proportions (out of complete) for each variable and cohort (same as `get_proportion()`)

\item \code{get_count_prop()}: Get a string formatted as "[count] ([prop])" for each response and cohort

\item \code{get_percent()}: Get the percents (out of complete) for each variable and cohort

\item \code{get_count_percent()}: Get a string formatted as "[count] ([percent]\%)" for each response and cohort

}}
\examples{
test_ct <- crosstab(length_by_species, "species")

# Base values
get_complete(test_ct)
get_mean(test_ct)
get_sd(test_ct)

# Formatted column
get_complete_total(test_ct)
get_med_q1_q3(test_ct)

# Raw joined values
get_complete_total(test_ct, raw = TRUE)
get_med_q1_q3(test_ct, raw = TRUE)

}
