% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vals.R
\name{join_val}
\alias{join_val}
\title{Join Multiple Value Results}
\usage{
join_val(...)
}
\arguments{
\item{...}{As many data frames as you wish to join, separated by commas}
}
\value{
The resulting full-joined data frame
}
\description{
Provides an easy interface to join the multiple variables you need into one
data frame. This is essentially [dplyr::full_join()] but automatically
detects which columns to join by and allows you to pass in as many tables as
you want to join.
}
\examples{
test_df <- iris[, c("Petal.Length", "Species"), drop = FALSE]
test_ct <- crosstab(test_df, "Species")

# One cohort and one variable column each
join_val(
    get_mean(test_ct),
    get_sd(test_ct),
    get_mean_sd(test_ct)
)

# Tables with multiple columns
join_val(
    get_complete_total(test_ct, raw = TRUE),
    get_med_q1_q3(test_ct, raw = TRUE)
)

}
