% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kable_extensions.R
\name{kbl}
\alias{kbl}
\title{Create Kable Tables from Crosstab Objects}
\usage{
kbl(x, crosstab_footnotes = TRUE, pack_rows = TRUE, table_sep = TRUE, ...)
}
\arguments{
\item{x}{The crosstab object, or other object to be passed to [kableExtra::kbl()]}

\item{crosstab_footnotes}{Logical - Add the auto-generated footnotes?}

\item{pack_rows}{Logical - Add the auto-generated row grouping?}

\item{table_sep}{Logical - Add the auto-generated lines separating stacked tables?}

\item{...}{All other arguments to be passed to [kableExtra::kbl()]}
}
\value{
The kable object returned from [kableExtra::kbl()]
}
\description{
`kbl()` is a wrapper for [kableExtra::kbl()] and can be used as a drop-in
replacement. If any object besides a crosstab is passed in, it will simply
pass the input arguments on to [kableExtra::kbl()]. If a crosstab is passed
in, it can automatically apply auto-generated footnotes, row-packing, and
table separators.
}
\details{
`kbl()` is a wrapper for [kableExtra::kbl()] and can be used as a drop-in
replacement. For crosstab objects, `kbl()` calls [apply_footnotes()],
[apply_pack_rows()], and [apply_table_sep()] unless otherwise specified in
the arguments.
}
\section{Superscripting}{

If the crosstab object has ANOVA markers with `superscript = TRUE` then
`escape` is set to `FALSE` before passing the arguments on to
[kableExtra::kbl()]. This is because the superscripted markers have <sup>
HTML tags. `kbl()` automatically escapes these, but we want the raw output
passed directly to LaTeX. The entire table then has all LaTeX special
characters manually escaped, which will cause issues in the Viewer window,
but not in the final output.

I could probably figure out how to make it work with the Viewer window, but
I'll be honest, I don't wanna. :)
}

\examples{
library(kableExtra, warn.conflicts = FALSE)

test_ct <- crosstab_stacked(
    iris,
    "Species",
    anova_format = c("row", "marker")
)

# Default formatting
htcrosstabs::kbl(test_ct)

# Use with other kableExtra formatting just like kableExtra::kbl()
test_ct |>
    htcrosstabs::kbl(
        booktabs = TRUE,
        align = c("l", "r", "r", "r", "r"),
        longtable = TRUE
    ) |>
    kableExtra::kable_styling(
        latex_options = c("striped", "repeat_header", "hold_position"),
        position = "center"
    )

}
