% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{length_by_species}
\alias{length_by_species}
\title{Petal Length by Species}
\format{
A data frame with 150 rows of iris petal lengths and species
\describe{
  \item{petal length}{Length of flower petals in cm (numeric)}
  \item{species}{Species of iris: virginica, versicolor, and setosa (factor)}
}
}
\source{
The famous "iris" dataset built into R
}
\usage{
length_by_species
}
\description{
The petal length and species columns from the famous "iris" dataset built into R.
}
\keyword{datasets}
