% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{licorice_by_region}
\alias{licorice_by_region}
\title{Opinions on Black Licorice by US Region}
\format{
A data frame with 400 rows of black licorice opinions and US region.
\describe{
  \item{opinion}{Opinion on black licorice (factor)}
  \item{region}{Region of the US: west, northeast, north central, and south (factor)}
}
}
\source{
Dummy data, invented by the developer
}
\usage{
licorice_by_region
}
\description{
A dummy dataset of Likert-like opinions on black licorice by US region (west,
northeast, north central, and south). The region breakdown is inspired by the
`state.region` dataset.
}
\keyword{datasets}
