% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{petal_width}
\alias{petal_width}
\title{Petal Width}
\format{
A data frame with 150 rows of iris petal widths
\describe{
  \item{petal width}{Width of flower petals in cm (numeric)}
}
}
\source{
The famous "iris" dataset built into R
}
\usage{
petal_width
}
\description{
The petal width column from the famous "iris" dataset built into R.
}
\keyword{datasets}
