% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{satisfaction_by_company}
\alias{satisfaction_by_company}
\title{Job Satisfaction by Company}
\format{
A data frame with 400 rows of satisfaction responses about working at various companies.
\describe{
  \item{satisfaction}{Job satisfaction (factor)}
  \item{company}{Tech company: Apple, Amazon, Google, and Microsoft (factor)}
}
}
\source{
Dummy data, invented by the developer
}
\usage{
satisfaction_by_company
}
\description{
A dummy dataset of Likert-like satisfaction responses about working at
different tech companies (Apple, Amazon, Google, and Microsoft).
}
\keyword{datasets}
