% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{sports_by_age}
\alias{sports_by_age}
\title{Sports by Age}
\format{
A data frame with 400 rows of preferred sports and age groups
\describe{
  \item{sport}{Preferred sport (factor)}
  \item{age}{Age group: child, teen, adult, or senior (factor)}
}
}
\source{
Dummy data, invented by the developer
}
\usage{
sports_by_age
}
\description{
A dummy dataset of preferred sports by age group (child, teen, adult, and
senior)
}
\keyword{datasets}
