% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosstab.R
\name{stack_crosstabs}
\alias{stack_crosstabs}
\title{Combine Multiple Crosstabs}
\usage{
stack_crosstabs(...)
}
\arguments{
\item{...}{As many crosstab objects as you want to combine, separated by
commas}
}
\value{
One crosstab with all rows from all crosstabs passed in, and the data table from the final crosstab
}
\description{
`stack_crosstabs()` takes crosstab objects and combines them into one
crosstab in the order that they are provided. The resulting crosstab's data
table will be the same as the final crosstab passed to this function.
}
\examples{
ct1 <- iris[, c("Sepal.Length", "Species"), drop = FALSE] |>
    crosstab(cohort_col_name = "Species") |>
    add_default_table()

ct2 <- iris[, c("Petal.Length", "Species"), drop = FALSE] |>
    crosstab(cohort_col_name = "Species") |>
    add_default_table()

ct3 <- iris[, c("Sepal.Width", "Species"), drop = FALSE] |>
    crosstab(cohort_col_name = "Species") |>
    add_default_table()

ct4 <- iris[, c("Petal.Width", "Species"), drop = FALSE] |>
    crosstab(cohort_col_name = "Species") |>
    add_default_table()

combined_ct <- stack_crosstabs(ct1, ct2, ct3, ct4)

combined_ct

}
