% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{students}
\alias{students}
\title{University Student Data}
\format{
A data frame with 3500 rows of student profile data.
\describe{
  \item{university}{The university name (factor)}
  \item{athletics}{If the student is involved in athletics (logical)}
  \item{height}{Height of the student in cm (numeric)}
  \item{semester_credits}{Credits enrolled in the current semester (integer)}
  \item{department}{Major department category (factor)}
  \item{uni_perception}{How this student feels about the university (factor)}
  \item{prof_support}{How supportive the students feel their average professor is (factor)}
  \item{food_court_per_week}{Which restaurants in the food courts they've visited this week}
}
}
\source{
Dummy data, invented by the developer
}
\usage{
students
}
\description{
A dummy dataset of various data types covering student profiles at well-known
universities: UCLA, Berkeley, Stanford, Yale, MIT, Texas Tech, and Alabama.
}
\keyword{datasets}
