% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rows.R
\name{to_long}
\alias{to_long}
\title{Pivot Rows Longer}
\usage{
to_long(wide_df, description_col, cohorts_to, values_to)
}
\arguments{
\item{wide_df}{The wide dataframe in the typical output format}

\item{description_col}{The name of the description column}

\item{cohorts_to}{The name of the cohort column to be created}

\item{values_to}{The name of the variable column to be created}
}
\value{
The data in long form, easier for extracting data from
}
\description{
Pivots data in wide form (for output) to a longer form easier to extract data
from after formatting. More specified in purpose than
`tidyr::pivot_longer()`.
}
\details{
#' The data passed into `wide_df` should contain:
* A description column
* Cohort columns
}
\examples{
wide_df <- data.frame(
    Description = c("Granny Smith", "Red Delicious"),
    All = c(58, 23),
    Children = c(12, 11),
    Teenagers = c(16, 8),
    Adults = c(30, 4)
)
wide_df

to_long(
    wide_df = wide_df,
    description_col = "Description",
    cohorts_to = "Age ranges",
    values_to = "Num favorite"
)
}
