% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rows.R
\name{to_wide}
\alias{to_wide}
\title{Pivot Rows Wider}
\usage{
to_wide(long_df, desc_col, cohort_col, na_fill = NA)
}
\arguments{
\item{long_df}{The data frame with the data in long form (3 columns: description, cohort, and variable)}

\item{desc_col}{The name of the description column (the far left in the output tables)}

\item{cohort_col}{The name of the cohort column}

\item{na_fill}{The values to replace NAs with when pivoting}
}
\value{
The data in wide form, formatted for the output table
}
\description{
Pivots data in long form to the wide format used in the output table.
More speficied in purpose than [tidyr::pivot_wider()] functions.
}
\details{
The data passed into `long_df` should contain three columns:
* A description column (the leftmost column in the output tables)
* A cohort column
* A variable column
}
\examples{
# Original long form
num_ct <- crosstab(length_by_species, "species")
means <- get_mean(num_ct)
means

# With description column
means[["Description"]] <- "Mean Petal Length"
means

# Wide for output
to_wide(means, "Description", "species")

}
