% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_casting.R
\name{check_crosstab}
\alias{check_crosstab}
\alias{is.crosstab}
\alias{is.crosstab.data}
\alias{is.crosstab.categorical}
\alias{is.crosstab.numeric}
\alias{is.crosstab.likert}
\alias{is.crosstab.multi}
\alias{is.crosstab.grouped}
\alias{assert_crosstab}
\alias{assert_crosstab_data}
\alias{assert_crosstab_categorical}
\alias{assert_crosstab_numeric}
\alias{assert_crosstab_likert}
\alias{assert_crosstab_multi}
\alias{assert_crosstab_grouped}
\title{Check or Assert Crosstab Object Types}
\usage{
is.crosstab(obj, strict = TRUE)

is.crosstab.data(obj)

is.crosstab.categorical(obj)

is.crosstab.numeric(obj)

is.crosstab.likert(obj)

is.crosstab.multi(obj)

is.crosstab.grouped(obj)

assert_crosstab(obj, strict = TRUE)

assert_crosstab_data(obj)

assert_crosstab_categorical(obj)

assert_crosstab_numeric(obj)

assert_crosstab_likert(obj)

assert_crosstab_multi(obj)

assert_crosstab_grouped(obj)
}
\arguments{
\item{obj}{An object to check.}

\item{strict}{Logical. If `TRUE`, only returns `TRUE` for fully constructed
crosstab objects (will return false for raw data subclasses).}
}
\value{
Logical for `is.*()` and `assert_*()` functions (`assert_*()` functions throw error if check fails).
}
\description{
`is.crosstab()` and related functions test or assert the structure of a
crosstab object and its subclasses (e.g., numeric, categorical, likert,
multi-response, grouped). return `TRUE` if the object is a valid crosstab
(or specific subtype). The `assert_` functions will throw an error
if the check fails.
}
\section{Functions}{
\itemize{
\item \code{is.crosstab()}: Checks if the data is a crosstab object

\item \code{is.crosstab.data()}: Checks if the data is specifically a crosstab_data object

\item \code{is.crosstab.categorical()}: Checks if the data is categorical

\item \code{is.crosstab.numeric()}: Checks if the data is numeric

\item \code{is.crosstab.likert()}: Checks if the data is Likert-like

\item \code{is.crosstab.multi()}: Checks if the data is multi-response

\item \code{is.crosstab.grouped()}: Checks if the data is grouped

\item \code{assert_crosstab()}: Throws error if the data is not a crosstab

\item \code{assert_crosstab_data()}: Throws error if the data is not a crosstab_data

\item \code{assert_crosstab_categorical()}: Throws error if the data is not categorical

\item \code{assert_crosstab_numeric()}: Throws error if the data is not numeric

\item \code{assert_crosstab_likert()}: Throws error if the data is not Likert-like

\item \code{assert_crosstab_multi()}: Throws error if the data is not multi-response

\item \code{assert_crosstab_grouped()}: Throws error if the data is not grouped

}}
\examples{
# Numeric, no grouping column
num_ct_ungrouped <- crosstab(petal_width)

is.crosstab(iris) # FALSE
is.crosstab(num_ct_ungrouped) # TRUE
is.crosstab.categorical(num_ct_ungrouped) # FALSE
is.crosstab.numeric(num_ct_ungrouped) # TRUE

# Numeric, grouped
num_ct_grouped <- crosstab(length_by_species, "species")

is.crosstab.grouped(num_ct_ungrouped) # FALSE
is.crosstab.grouped(num_ct_grouped) # TRUE

# Differentiating complete crosstabs from inner data objects
num_ct_data <- crosstab_data(length_by_species, "species")

is.crosstab.data(num_ct_ungrouped) # FALSE
is.crosstab.data(num_ct_data) # TRUE
is.crosstab(num_ct_data) # FALSE (only fully constructed crosstabs return true)
is.crosstab(num_ct_data, strict = FALSE) # TRUE (any part of a crosstab returns TRUE)
}
