% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{nest_multi_col}
\alias{nest_multi_col}
\title{Nest Column to List of Atomic Vectors}
\usage{
nest_multi_col(df, multi_col_name)
}
\arguments{
\item{df}{A data frame to nest}

\item{multi_col_name}{The name of the column to nest}
}
\value{
A data frame with nested atomic vectors in the specified column
}
\description{
[tidyr::nest()] Condenses rows into lists of tibbles, but we need a list of
atomic vectors. This function does just that. It groups by every other column
and then collapses the duplicates.
}
\examples{
# Create unnested data
unnested_df <- allergies_by_school
unnested_df[["id"]] <- 1:nrow(unnested_df)
unnested_df <- tidyr::unnest(unnested_df, allergies)
head(unnested_df, 5)

# Nest data
nested_df <- nest_multi_col(unnested_df, "allergies")
head(nested_df, 5)

}
